Move data folder into public folder
