--- src/data/README.md
@@ -1,18 +1,0 @@
-# parse.py
-
-This script is used to parse the .tsv file and convert it to .csv file.
-
-Usage: 
-- `python parse.py <input_tsv_file> <output_csv_file>`
-- `python parse.py <input_tsv_file>`
-- `python parse.py <name_of_tsv_file> <name_of_csv_file>`
-- `python parse.py <name_of_tsv_file>`
-
-## Parsed Files
-| Index | Title                                                         | Parsed | Parsing Date | Parsing Time | Notes            |
-| ----: | :------------------------------------------------------------ | :----: | :----------: | :----------: | :--------------- |
-| 0     | [EVDa_SimCaTip_Ale0003](csv/EVDa_SimCaTip_Ale0003.csv)        | Yes    | 02/15/2024   | 13:08        | Don't use        |
-| 1     | [EVDb_SimCaPlus_Mario0006](csv/EVDb_SimCaPlus_Mario0006.csv)  | Yes    | 02/17/2024   | 11:23        |                  |
-| 3     | [EVDb_SimCaPlus_Ale0005](csv/EVDb_SimCaPlus_Ale0005.csv)      | Yes    | 02/17/2024   | 11:23        |                  |
-| 4     | [EVDb_SimCaPlus_Ale0004](csv/EVDb_SimCaPlus_Ale0004.csv)      | Yes    | 02/19/2024   | 21:45        |                  |
-| 5     | [EVDb_SimCaPlus_Mario0007](csv/EVDb_SimCaPlus_Mario0007.csv)  | Yes    | 02/19/2024   | 21:45        |                  |
